#162 separando aplicação da conexão com banco de dados e server.
